requirements.txt
Django==3.2.5
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-resized==0.3.11
